Refactor document chunker package
