+++ src/document_chunker/document_chunker.py
@@ -1,0 +1,106 @@
+import logging
+from dataclasses import dataclass
+from enum import Enum
+from typing import Dict, List, Optional, Set
+
+from textractor.entities.document import Document
+
+from data_models.chunk_models import DocumentMetadata, OpenSearchChunk
+from document_chunker.chunking_strategies import ChunkingStrategyHandler, LineBasedChunkingHandler
+
+logger = logging.getLogger(__name__)
+
+
+class ChunkingStrategy(Enum):
+    """Available chunking strategies."""
+
+    LINE_BASED = "line_based"
+    # SENTENCE_BASED = "sentence_based"
+    # TOKEN_BASED = "token_based"
+
+
+@dataclass
+class ChunkingConfig:
+    """Configuration for chunking behavior."""
+
+    maximum_chunk_size: int = 1000
+    strategy: ChunkingStrategy = ChunkingStrategy.LINE_BASED
+
+
+class DocumentChunker:
+    """Handles extraction of chunks from Textractor documents."""
+
+    def __init__(self, config: Optional[ChunkingConfig] = None):
+        self.config = config or ChunkingConfig()
+
+        self.strategy_handlers: Dict[str, ChunkingStrategyHandler] = {
+            "LAYOUT_TEXT": LineBasedChunkingHandler(self.config.maximum_chunk_size),
+            # "LAYOUT_TABLE": TableChunkingHandler(self.config.maximum_chunk_size), # Future
+            # "LAYOUT_LIST": ListChunkingHandler(self.config.maximum_chunk_size),   # Future
+        }
+        # A default handler for any types not explicitly mapped
+        self.default_handler = LineBasedChunkingHandler(self.config.maximum_chunk_size)
+
+    def chunk(
+        self, doc: Document, metadata: DocumentMetadata, desired_layout_types: Optional[Set[str]] = None
+    ) -> List[OpenSearchChunk]:
+        """
+        Parses a Textractor Document and extracts specified layout blocks as structured chunks.
+
+        Args:
+            doc: Textractor Document to process
+            metadata: Document metadata
+            desired_layout_types: Set of layout types to extract (defaults to {"LAYOUT_TEXT"})
+
+        Returns:
+            List of OpenSearchChunk objects
+
+        Raises:
+            ValueError: If metadata validation fails
+        """
+        try:
+            self._validate_inputs(doc, metadata)
+
+            if desired_layout_types is None:
+                desired_layout_types = {"LAYOUT_TEXT"}
+
+            chunks = []
+            chunk_index_counter = 0
+
+            for page in doc.pages:
+                page_chunks = self._process_page(page, metadata, desired_layout_types, chunk_index_counter)
+                chunks.extend(page_chunks)
+                chunk_index_counter += len(page_chunks)
+
+            logger.info(f"Extracted {len(chunks)} chunks from document {metadata.ingested_doc_id}")
+            return chunks
+
+        except Exception as e:
+            logger.error(f"Error extracting chunks from document {metadata.ingested_doc_id}: {str(e)}")
+            raise
+
+    def _validate_inputs(self, doc: Document, metadata: DocumentMetadata) -> None:
+        """Validate inputs before processing."""
+        if not doc or not doc.pages:
+            raise ValueError("Document cannot be None and must contain pages.")
+
+    def _process_page(
+        self, page, metadata: DocumentMetadata, desired_layout_types: Set[str], chunk_index_start: int
+    ) -> List[OpenSearchChunk]:
+        """Process a single page and return its chunks."""
+        chunks = []
+        chunk_index = chunk_index_start
+
+        for layout_block in page.layouts:
+            if self._should_process_block(layout_block, desired_layout_types):
+                block_type = layout_block.layout_type
+                handler_to_use = self.strategy_handlers.get(block_type, self.default_handler)
+                block_chunks = handler_to_use.chunk(layout_block, page, metadata, chunk_index)
+                chunks.extend(block_chunks)
+                chunk_index += len(block_chunks)
+
+        return chunks
+
+    def _should_process_block(self, layout_block, desired_layout_types: Set[str]) -> bool:
+        """Determine if a layout block should be processed."""
+        return layout_block.layout_type in desired_layout_types and layout_block.text and layout_block.text.strip()
