+++ src/document_chunker/tests/data/textract_response_template.json
